Added Token Blacklisting
The idea is very simple but smart. the main reason of using this approach is to avoid using sessions. it basically blocks any token that was issued before a certian point in time. ex: any token that was issued before the time of changing a password will no longer be valid.